Common knn graph build params (#949)
Reusing graph build params for cagra and all neighbors.
Related issue: https://github.com/rapidsai/cuvs/issues/931

Authors:
  - Jinsol Park (https://github.com/jinsolp)
  - Corey J. Nolet (https://github.com/cjnolet)

Approvers:
  - Micka (https://github.com/lowener)

URL: https://github.com/rapidsai/cuvs/pull/949
